* commit '3c5eafdd10a463b7c137101a83e1d78ee66d4341':
  No reason to send in the user for this method.
  Added: New api command to fetch user permissions Refactoring:  Pulled out argument creation, made repo name optional

+++ stash_client/rest.py
@@ -4,6 +4,7 @@ REST client for Stash.
 import requests.api
 import json
 import logging
+import os
 from ConfigParser import SafeConfigParser
 
 STASH_HOST = None
@@ -15,6 +16,7 @@ BASE_URL = 'https://%s/rest/api/%s' % (STASH_HOST, STASH_API_VERSION)
 _PROJECT_NAMESPACE = 'projects'
 _USER_NAMESPACE = 'users'
 _REPOSITORY_NAMESPACE = 'repos'
+_PERMISSIONS = 'permissions'
 
 
 class UserError(Exception):
@@ -30,11 +32,16 @@ class ResponseError(Exception):
             self.response.status_code, self.response.reason)
 
 
-def set_creds(username, password):
+def set_creds(args):
+    import getpass
+    
+    username = os.environ["USER"]
+    if args.user_override:
+        username = args.user_override
+    password = getpass.getpass("Stash password for %s: " % username)
     global USERNAME
     global PASSWORD
     (USERNAME, PASSWORD) = (username, password)
-
 
 def create_url(user=None, project=None, repository=None, api_path=None):
     if user is not None and project is not None:
@@ -66,30 +73,52 @@ def post_json(user=None, project=None, repository=None, api_path=None, post_data
         raise ResponseError(resp)
     return resp
 
+def get_json(user=None, project=None, repository=None, api_path=None, post_data=None):
+    if post_data is None:
+        raise Exception("POST data is not actually allowed to be None")
+    api_url = create_url(user=user, project=project, repository=repository, api_path=api_path)
+    json_string = json.dumps(post_data)
+    logging.debug("Posting %s to %s", json_string, api_url)
+    resp = requests.api.get(api_url, data=json_string, auth=(USERNAME, PASSWORD),
+                             headers={'Content-type': 'application/json'})
+    if not resp.ok:
+        raise ResponseError(resp)
+    return resp
 
 def create_repo(repository_name, user=None, project=None):
+    if(repository_name is None):
+        raise UserError("You must specify a repository")
     if user is None and project is None:
         raise UserError("new repository needs a project or a user")
     post_data = {'name': repository_name}
     return post_json(post_data=post_data, user=user, project=project, api_path=[_REPOSITORY_NAMESPACE])
 
-
+def list_user_permissions( user=None, project=None, group_name=None):
+    if project is None:
+        raise UserError("Listing project permisisons needs a project")
+    api_cmd = _PERMISSIONS + "/users"
+    post_data = {'filter':group_name}
+    resp = get_json(post_data=post_data, user=user, project=project, api_path=[api_cmd])
+    if resp.text:
+        values = resp.json()["values"]
+        for value in values :
+            print value["user"]["displayName"] + " : " + value["permission"]
+    else:
+        raise ResponseError("List users attempt failed with status %d: %s" % (resp.status_code, resp.reason))
+    
 def delete_repo(repository_name, user=None, project=None):
+    if(repository_name is None):
+        raise UserError("You must specify a repository")
     if user is None and project is None:
-        raise UserError("deleted repository needs a project or a user")
+        raise UserError("deleting a repository needs a project or a user")
     api_url = create_url(user=user, project=project, repository=repository_name)
     resp = requests.api.delete(api_url, auth=(USERNAME, PASSWORD))
     if not resp.ok:
         raise ResponseError(resp)
     return resp
 
-
-def main():
-    import getpass
+def get_cmd_arguments():
     from argparse import ArgumentParser
-    import os
-    from .models import StashRepo
-    logging.basicConfig()
     parser = ArgumentParser()
     parser.add_argument("-o", "--organization", action="store", dest="org",
                         help=("Github organization or Stash project to query for repositories to clone."
@@ -98,11 +127,21 @@ def main():
                         help="User to query for repositories to clone.  Either this or -o is required.")
     parser.add_argument("-U", "--override_user", action="store", dest="user_override",
                         help="Override the local user for accessing stash.  If not specified, local user will be used.")
+    parser.add_argument("-C", "--create", action="store_true", dest="create",
+                        help="Create a repo")
+    parser.add_argument("-perm", "--list_user_permissions", action="store_true", dest="list_user_permissions",
+                        help="List the permissions for the users of this project")
     parser.add_argument("-D", action="store_true", dest="delete",
                         help="DELETE IT.")
     parser.add_argument("-v", "--verbose", action="store_true", dest="verbose", help="Log INFO to STDOUT")
-    parser.add_argument("repo_name", nargs=1)
-    args = parser.parse_args()
+    parser.add_argument("-r", "--repo_name", action="store", dest="repo_name", help="The name of the repository.")
+    return parser.parse_args()
+    
+def main():
+    from .models import StashRepo
+    logging.basicConfig()
+    args = get_cmd_arguments()
+    
     if args.verbose:
         logging.getLogger().setLevel(logging.DEBUG)
     # silly approach that avoids hard-coding the stash repo
@@ -111,22 +150,23 @@ def main():
     global STASH_HOST
     STASH_HOST = config.get('server', 'hostname')
 
-    username = os.environ["USER"]
-    if args.user_override:
-        username = args.user_override
-    password = getpass.getpass("Stash password for %s: " % username)
-    set_creds(username, password)
-
     if args.delete:
-        resp = delete_repo(args.repo_name[0], user=args.user, project=args.org)
+        set_creds(args)
+        resp = delete_repo(args.repo_name, user=args.user, project=args.org)
         if resp.text:
             print "Deletion OK: %s" % resp.json().get('message')
         else:
             print "Deletion attempt succeeded with status %d: %s" % (resp.status_code, resp.reason)
-    else:
-        resp = create_repo(args.repo_name[0], user=args.user, project=args.org)
+    elif args.create:
+        set_creds(args)
+        resp = create_repo(args.repo_name, user=args.user, project=args.org)
         repo = StashRepo(resp.json())
         print "Successfully created repo %s with clone URL %s" % (repo.name, repo.get_clone_url('ssh'))
+    elif args.list_user_permissions:
+        set_creds(args)
+        list_user_permissions(project=args.org)
+    else:
+        print "No operation specified."
 
 
 if '__main__' == __name__:
